requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.2
h11==0.14.0
idna==3.4
pip==22.3.1
pydantic==1.10.7
pypng==0.20220715.0
setuptools==65.5.1
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.6.0
wheel==0.38.4
pymongo~=4.3.3
python-dotenv==1.0.0